move store declarations to proper folder
